requirements.txt
fastapi==0.75.0
pandas==1.4.1
pydantic==1.9.0
pymongo==4.0.2
python-dotenv==0.19.2
requests==2.22.0
uvicorn==0.17.6
google-api-python-client
google-auth
numpy
